requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
lxml==6.0.1
python-dateutil==2.9.0.post0
pyvmomi==9.0.0.0
pyyaml==6.0.2
requests==2.32.5
six==1.17.0
tqdm==4.67.1
urllib3==2.5.0
webdavclient3==3.14.6
